requirements.txt
scapy>=2.5.0
PyYAML>=6.0
pytest>=6.0
